refactor environment
